Refactor functionality of number and operation buttons
The number and operation buttons' functionalities
have been separated into two event handler blocks.
This is because whenever a calculation is made, the
first keystroke should not be an operator but a number.
The problem with the original functionality (one
event handler that gives the same functionality to
both the number and operation buttons) is that it
is hard to distinguish numbers from operations at
the start of a calculation without creating complexity
in the program. Separating the buttons' functionalities
into multiple event handlers solves this.