Refactor method of displaying calculations
The previous method of displaying calculations was
to have all of the keystrokes appear in the answer
box and the method of calculation was to take the
string in the answer box, split it, and operate on
it. However, this makes it difficult to string
operations together because the string in the answer
box would need to be split by its operator multiple
times across the operands. Showing each number as
it is clicked and highlighting the operator clicked
between calculations takes away the need for
splitting the string and gives more flexibility in
calculating subsequent equations.