Refactor functionality to allow stringing operations
Before, the goal of the calculator was to perform
single calculations and when the result was displayed
using the equals button, a new operand could be keyed
in by the user to start a new single calculation. This
was possible by the presence of a conditional variable
isNewCalculation, which had a start value of false and
flipped to true when the equals button was clicked
(a calculation had occurred), allowing it to clear
the answer box when a new operand was keyed in. It
has been removed, as the new stringing functionality
of the calculator allows for this without the variable
by seeing which key is keyed in next by the user.
If the next key (after the equals button is clicked)
is a number, it must be a new calculation; otherwise
it is a stringing operation.