Refactor code to allow for stringing via op buttons
Before, the stringing only worked once a calculation
was completed, via clicking the equals button. This
new rendition allows for stringing via clicking on
any of the operation buttons after two operands and
an operator have been keyed in (ex. if you key in
1, then +, then 2, pressing any of the other op
buttons after will result in 3 appearing in the
answer box) by transferring all the code that runs
after the equals button is clicked into a function
displayResult and calling it in the operation buttons'
event handler.