Fix issues with delete button between calculations
The calculator was computing calculations based on
numbers that was specified in the code and not the
numbers on the screen. This was because the delete
button was not properly updating the operands' values
after removing their digits in the answer box. This
rendition corrects these issues with both the first
and second operands in each calculation.