Fix issues with deleting multiple digits
The last rendition only accounted for updates when
only one digit is removed from the first operand.
This new rendition fixes the issue by leveraging
that the second operand does not exist at this point
in the calculation in order to distinguish itself as
the first operand of the calculation.